+++ NEWS.md
@@ -1,5 +1,9 @@
 # CytoPipelineUtils 0.99
 
+## CytoPipelineUtils 0.99.3
+- applyFlowJoGate() can now match sample file with sample names in wsp file,
+by using pattern matching, instead of full file name matching
+
 ## CytoPipelineUtils 0.99.2
 - added 2nd version of removeDebrisFlowClust()
 - applyFlowJoGate() can now use different flowJo wsp files, depending on pData